+++ src/ChatServer/GroupMessages.cpp
@@ -29,7 +29,7 @@ void ChatMessageLib::sendIsmInviteRequest(Player* sender, Player* target) const
 	gMessageFactory->addUint64(sender->getGroupId()); // the group id
 	newMessage = gMessageFactory->EndMessage();
 
-	target->getClient()->SendChannelA(newMessage,target->getClient()->getAccountId(),target->getPlanetId()+8, 3);
+	target->getClient()->SendChannelA(newMessage,target->getClient()->getAccountId(),static_cast<uint8>(target->getPlanetId())+8, 3);
 }
 
 //======================================================================================================================
@@ -42,7 +42,7 @@ void ChatMessageLib::sendIsmGroupLootModeResponse(Player* target) const
 	gMessageFactory->addUint32(opIsmGroupLootModeResponse);
 	gMessageFactory->addUint32(target->getClient()->getAccountId());  // the sender's id
 	newMessage = gMessageFactory->EndMessage();
-	target->getClient()->SendChannelA(newMessage,target->getClient()->getAccountId(),target->getPlanetId()+8, 3);
+	target->getClient()->SendChannelA(newMessage,target->getClient()->getAccountId(),static_cast<uint8>(target->getPlanetId())+8, 3);
 }
 
 //======================================================================================================================
@@ -55,7 +55,7 @@ void ChatMessageLib::sendIsmGroupLootMasterResponse(Player* target) const
 	gMessageFactory->addUint32(opIsmGroupLootMasterResponse);
 	gMessageFactory->addUint32(target->getClient()->getAccountId());  // the sender's id
 	newMessage = gMessageFactory->EndMessage();
-	target->getClient()->SendChannelA(newMessage,target->getClient()->getAccountId(),target->getPlanetId()+8, 3);
+	target->getClient()->SendChannelA(newMessage,target->getClient()->getAccountId(),static_cast<uint8>(target->getPlanetId())+8, 3);
 }
 
 //======================================================================================================================
@@ -73,7 +73,7 @@ void ChatMessageLib::sendIsmGroupCREO6deltaGroupId(uint64 groupId, Player* targe
 	gMessageFactory->addUint64(groupId); // the new group id
 	newMessage = gMessageFactory->EndMessage();
 	
-	target->getClient()->SendChannelA(newMessage,target->getClient()->getAccountId(),target->getPlanetId()+8, 3);
+	target->getClient()->SendChannelA(newMessage,target->getClient()->getAccountId(),static_cast<uint8>(target->getPlanetId())+8, 3);
 }
 
 //======================================================================================================================
